bench: promote five synth fixtures out of _pending
`check.py --synthetic-only --synthetic-pattern '_pending/*.py'` passes these on
dynasm, cranelift and wasm, so they now run in the default gate:

  exception_nested_exc_info_restore    JIT answered 3320000 (trait) / 3360000
                                       (FBW walker); now 360000, matching the
                                       interpreter and CPython
  gc_bug_bridge_flavor_traceback_names cranelift aborted 8/10 to 9/10 runs with
                                       `GC BUG: invalid type_id=` out of
                                       `incremental_mark_step`; now 20/20 clean
                                       with the output its header pins
  loop_callee_shared_mutation          JIT printed 2*N + 3; now 2*N
  sre_wasm_min, sre_wasm_min1          pass on all three backends

Their headers described the failures as current, so each is restated as a
regression oracle keeping what the defect was.

The remaining four stay in `_pending/`, none of them for JIT reasons:
`dict_set_clear_in_eq_ops` and `dict_set_clear_in_eq_restart` record a
deliberate CPython/PyPy divergence, which check.py reports as BASEFAIL by
construction; `wasm_parse_template_int_ref` imports `re._parser`, absent before
CPython 3.11, and the local baseline is 3.9.6; `cranelift_finalizer_ordering_crash`
exceeds the 20s per-benchmark timeout.

check.py: dynasm 333/333, cranelift 333/333, wasm 330/330.
