+++ pyre/bench/synth/sre_wasm_min.py
