+++ pyre/bench/synth/sre_wasm_min1.py
